updated handlers TODO implement prime graphs
